# remove the initialization of saved_mean and saved_variance for batch_norm op